fix: hydration error preventing level loading
